fetch config file synchronously at startup
