--- src/config/fetch-config.d.ts
@@ -1,5 +1,0 @@
-import { AppConfig } from "interfaces";
-
-declare async function getAppConfig(configFileSetting: string): AppConfig;
-
-export default getAppConfig;
