--- src/config/load-config.d.ts
@@ -1,4 +1,0 @@
-import { AppConfig } from "interfaces";
-
-declare const webpackLoadedConfig: AppConfig;
-export default webpackLoadedConfig;
